+++ tests/integration_tests/test_uds_api.py
@@ -35,7 +35,13 @@ class TestCumulusCreateCollectionDapa(TestCase):
             'c_data2': {'type': 'boolean'},
             'c_data3': {'type': 'keyword'},
         }
+
+        self.tenant = 'UDS_LOCAL_TEST'  # 'uds_local_test'  # 'uds_sandbox'
+        self.tenant_venue = 'DEV'  # 'DEV1'  # 'dev'
+        self.collection_name = 'SNDR-SNPP_ATMS@L1B$OUTPUT'  # 'uds_collection'  # 'sbx_collection'
+        self.collection_version = '24.03.20.14.40'.replace('.', '')  # '2402011200'
         return
+
 
     def test_add_admin_01(self):
         collection_url = f'{self.uds_url}admin/auth'
@@ -131,6 +137,53 @@ class TestCumulusCreateCollectionDapa(TestCase):
             self.assertTrue(v.startswith(self.uds_url), f'missing stage: {self.stage} in {v} for {k}')
         return
 
+    def test_catalog_get(self):
+        post_url = f'{self.uds_url}catalog'  # MCP Dev
+        # post_url = 'https://58nbcawrvb.execute-api.us-west-2.amazonaws.com/test/am-uds-dapa/collections/'  # MCP Dev
+        headers = {
+            'Authorization': f'Bearer {self.bearer_token}',
+        }
+        post_url = f'{post_url}?limit=100'
+        print(post_url)
+        query_result = requests.get(url=post_url,
+                                    headers=headers,
+                                    )
+        self.assertEqual(query_result.status_code, 200, f'wrong status code. {query_result.text}')
+        query_result = json.loads(query_result.text)
+        print(query_result)
+        self.assertEqual('Catalog', query_result['type'], f'wrong type: {query_result}')
+        self.assertTrue('links' in query_result, 'links missing')
+        links = {k['rel']: k for k in query_result['links']}
+        self.assertTrue('child' in links, f'missing next in links: {links}')
+        return
+
+    def test_collections_get_single_granule(self):
+
+        temp_collection_id = f'URN:NASA:UNITY:{self.tenant}:{self.tenant_venue}:{self.collection_name}___{self.collection_version}'
+
+        post_url = f'{self.uds_url}collections/{temp_collection_id}'  # MCP Dev
+        # post_url = 'https://58nbcawrvb.execute-api.us-west-2.amazonaws.com/test/am-uds-dapa/collections/'  # MCP Dev
+        headers = {
+            'Authorization': f'Bearer {self.bearer_token}',
+        }
+        # post_url = f'{post_url}?limit=100'
+        print(post_url)
+        query_result = requests.get(url=post_url,
+                                    headers=headers,
+                                    )
+        self.assertEqual(query_result.status_code, 200, f'wrong status code. {query_result.text}')
+        query_result = json.loads(query_result.text)
+        """
+        {'type': 'Collection', 'id': 'URN:NASA:UNITY:UDS_LOCAL_TEST:DEV:UDS_COLLECTION___2402011700', 'stac_version': '1.0.0', 'description': 'TODO', 'links': [{'rel': 'root', 'href': './collection.json?bucket=unknown_bucket&regex=%5Eabcd.1234.efgh.test_file.%2A%5C.data.stac.json%24', 'type': 'application/json', 'title': 'abcd.1234.efgh.test_file05.data.stac.json'}, {'rel': 'item', 'href': './collection.json?bucket=protected&regex=%5Eabcd.1234.efgh.test_file.%2A%5C.nc%24', 'type': 'data', 'title': 'abcd.1234.efgh.test_file05.nc'}, {'rel': 'item', 'href': './collection.json?bucket=protected&regex=%5Eabcd.1234.efgh.test_file.%2A%5C.nc.cas%24', 'type': 'metadata', 'title': 'abcd.1234.efgh.test_file05.nc.cas'}, {'rel': 'item', 'href': './collection.json?bucket=protected&regex=%5Eabcd.1234.efgh.test_file.%2A%5C.nc.cmr.xml%24', 'type': 'metadata', 'title': 'abcd.1234.efgh.test_file05.nc.cmr.xml'}, {'rel': 'item', 'href': './collection.json?bucket=protected&regex=%5Eabcd.1234.efgh.test_file.%2A%5C.nc.stac.json%24', 'type': 'metadata', 'title': 'abcd.1234.efgh.test_file05.nc.stac.json'}, {'rel': 'items', 'href': 'https://dxebrgu0bc9w7.cloudfront.net/collections/URN:NASA:UNITY:UDS_LOCAL_TEST:DEV:UDS_COLLECTION___2402011700/items', 'type': 'application/json', 'title': 'URN:NASA:UNITY:UDS_LOCAL_TEST:DEV:UDS_COLLECTION___2402011700 Granules'}], 'extent': {'spatial': {'bbox': [[0.0, 0.0, 0.0, 0.0]]}, 'temporal': {'interval': [['1970-01-01T12:00:00Z', '2024-02-26T07:11:11Z']]}}, 'license': 'proprietary', 'summaries': {'updated': ['2024-02-01T17:55:34.338000Z'], 'granuleId': ['^abcd.1234.efgh.test_file.*$'], 'granuleIdExtraction': ['(^abcd.1234.efgh.test_file.*)(\\.data\\.stac\\.json|\\.nc\\.cas|\\.cmr\\.xml)'], 'process': ['stac'], 'totalGranules': [1]}}
+        """
+        print(json.dumps(query_result, indent=4))
+        self.assertEqual('Collection', query_result['type'], f'wrong type: {query_result}')
+        self.assertEqual(temp_collection_id, query_result['id'], f'wrong collection_id: {query_result}')
+        self.assertTrue('links' in query_result, 'links missing')
+        links = {k['rel']: k for k in query_result['links']}
+        self.assertTrue('items' in links, f'missing items in links: {links}')
+        return
+
     def test_granules_get(self):
         post_url = f'{self.uds_url}collections/URN:NASA:UNITY:UDS_LOCAL_TEST:DEV:UDS_COLLECTION___2312041030/items/'  # MCP Dev
         headers = {
@@ -146,6 +199,21 @@ class TestCumulusCreateCollectionDapa(TestCase):
         links = {k['rel']: k['href'] for k in response_json['links'] if k['rel'] != 'root'}
         for k, v in links.items():
             self.assertTrue(v.startswith(self.uds_url), f'missing stage: {self.stage} in {v} for {k}')
+        return
+
+    def test_single_granule_get(self):
+        post_url = f'{self.uds_url}collections/URN:NASA:UNITY:UDS_LOCAL_TEST:DEV:UDS_COLLECTION___2312041030/items/URN:NASA:UNITY:UDS_LOCAL_TEST:DEV:UDS_COLLECTION___2312041030:test_file01'
+        headers = {
+            'Authorization': f'Bearer {self.bearer_token}',
+        }
+        print(post_url)
+        query_result = requests.get(url=post_url,
+                                    headers=headers,
+                                    )
+        self.assertEqual(query_result.status_code, 200, f'wrong status code. {query_result.text}')
+        response_json = json.loads(query_result.text)
+        print(json.dumps(response_json))
+
         return
 
     def test_create_new_collection(self):
@@ -183,9 +251,8 @@ class TestCumulusCreateCollectionDapa(TestCase):
         collection_created_result = requests.get(url=f'{post_url}{temp_collection_id}', headers=headers)
         self.assertEqual(collection_created_result.status_code, 200, f'wrong status code. {collection_created_result.text}')
         collection_created_result = json.loads(collection_created_result.text)
-        self.assertTrue('features' in collection_created_result, f'features not in collection_created_result: {collection_created_result}')
-        self.assertEqual(len(collection_created_result['features']), 1, f'wrong length: {collection_created_result}')
-        self.assertEqual(collection_created_result['features'][0]['id'], temp_collection_id, f'wrong id')
+        self.assertTrue('id' in collection_created_result, f'id not in collection_created_result: {collection_created_result}')
+        self.assertEqual(collection_created_result['id'], temp_collection_id, f'wrong id')
         print(collection_created_result)
         # TODO check if collection shows up
         return