+++ tf-module/unity-cumulus/any_to_lambda_module/any_to_lambda_variables.tf
@@ -1,0 +1,17 @@
+
+variable "rest_api_id" {
+  type = string
+}
+
+variable "resource_id" {
+  type = string
+}
+
+variable "authorizer_id" {
+  type = string
+}
+
+variable "lambda_invoke_arn" {
+  type = string
+}
+