+++ tf-module/unity-cumulus/variables.tf
@@ -68,6 +68,12 @@ variable "unity_default_provider" {
 
 }
 
+variable "cors_origins" {
+  default = ""
+  type = string
+  description = "Comma separated origins for CORS"
+}
+
 variable "dapa_api_prefix" {
   type = string
   description = "An API Gateway resource to identify the Project Name that this specific resource is integrated with"
@@ -135,4 +141,28 @@ variable "rest_api_stage" {
 variable "unity_cognito_authorizer__authorizer_id" {
   type = string
   description = "Example: 0h9egs"
+}
+
+variable "cors_200_response_parameters" {
+  type = map(bool)
+  default = {
+        "method.response.header.Access-Control-Allow-Credentials" = true
+        "method.response.header.Access-Control-Allow-Headers" = true
+        "method.response.header.Access-Control-Allow-Methods" = true
+        "method.response.header.Access-Control-Allow-Origin" = true
+        "method.response.header.Access-Control-Expose-Headers" = true
+        "method.response.header.Access-Control-Max-Age" = true
+    }
+}
+
+variable "cors_integration_response" {
+  type = map(string)
+  default = {
+        "method.response.header.Access-Control-Allow-Credentials" = "'true'",
+        "method.response.header.Access-Control-Allow-Headers" = "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
+        "method.response.header.Access-Control-Allow-Methods" = "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'",
+        "method.response.header.Access-Control-Allow-Origin" = "'*'"
+        "method.response.header.Access-Control-Expose-Headers" = "'Access-Control-Allow-Methods,Access-Control-Expose-Headers,Access-Control-Max-Age'"
+        "method.response.header.Access-Control-Max-Age" = "'300'"
+    }
 }