+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.2.0] - 2024-04-03
+### Changed
+- [#344](https://github.com/unity-sds/unity-data-services/pull/344) feat: update stac catalog list
+
+## [7.1.1] - 2024-04-02
+### Fixed
+- [#342](https://github.com/unity-sds/unity-data-services/pull/342) fix: no filename in main data.stac.json file
+
+## [7.1.0] - 2024-03-28
+### Changed
+- [#339](https://github.com/unity-sds/unity-data-services/pull/339) feat: testing stac browser ui
+
+## [7.0.2] - 2024-03-11
+### Fixed
+- [#337](https://github.com/unity-sds/unity-data-services/pull/337) fix: add cors to fast api
+
+## [7.0.1] - 2024-03-01
+### Fixed
+- [#335](https://github.com/unity-sds/unity-data-services/pull/335) fix: split into different catalog base url
+
+## [7.0.0] - 2024-02-29
+### Added
+- [#333](https://github.com/unity-sds/unity-data-services/pull/333) breaking: catalog & collection endpoint + items link
+
 ## [6.4.3] - 2024-02-06
 ### Fixed
 - [#325](https://github.com/unity-sds/unity-data-services/pull/325) fix: replace anything not alphanumeric to 3 underscores